refactor(lexer): move `misc_pre` and `misc_post` into separate module (#26482)
Pure refactor. No code changes, just moving code around.

`misc_pre` and `misc_post` were part of `classify` stage. Move them into their own separate stage.

Splitting it up into `pre.rs` and `post.rs` files is likely overkill - they could live within a single file. It's just a bit tidier for each stage to have its own directory (following the pattern of `classify`, `carve`, `coalesce` and `compress`) and it'd be weird if that directory just contained a single `mod.rs` file.

Splitting it does have the advantage of making it clear which primitives each uses from the `use` statements at top of each file.
