+++ crates/oxc_lexer/src/pipeline/misc/mod.rs
@@ -1,0 +1,4 @@
+mod post;
+mod pre;
+pub use post::misc_post;
+pub use pre::misc_pre;
